% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add_credentials}
\alias{add_credentials}
\title{Add credentials override for the right datasources}
\usage{
add_credentials(auth_type, config = list())
}
\arguments{
\item{auth_type}{Datasource Authentication type}

\item{config}{existing config override to extend}
}
\value{
named list of override configuration values
}
\description{
Add credentials override for the right datasources
}
