% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add_override}
\alias{add_override}
\title{Add override configuration values}
\usage{
add_override(..., .override = character())
}
\arguments{
\item{...}{named configuration values.}

\item{.override}{a named character vector}
}
\value{
named list of override configuration values
}
\description{
Add override configuration values
}
