% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_auth_type}
\alias{get_auth_type}
\title{Get datasource auth type (backward compatible)}
\usage{
get_auth_type(datasource)
}
\arguments{
\item{datasource}{datasource to retrieve auth type for}
}
\value{
Auth type of datasource
}
\description{
Get datasource auth type (backward compatible)
}
