% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore_functions.R
\name{get_object}
\alias{get_object}
\title{Retrieve an object from a datasource}
\usage{
get_object(client, datasource, object, override = new.env(), as = "raw")
}
\arguments{
\item{client}{\code{domino_data.data_sources.DataSourceClient}, as returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{override}{An environment with configuration overrides}

\item{as}{Passed through to \code{httr::content}}
}
\value{
Raw vector representation of the object
}
\description{
Retrieve an object from a datasource
}
