% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_aws_credentials}
\alias{load_aws_credentials}
\title{Load AWS Credentials}
\usage{
load_aws_credentials(location, profile = NULL)
}
\arguments{
\item{location}{file path where AWS credentials are located}

\item{profile}{AWS profile or section to use}
}
\value{
named list of override configuration values
}
\description{
Load AWS Credentials
}
