% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_oauth_credentials}
\alias{load_oauth_credentials}
\title{Load OAuth Credentials}
\usage{
load_oauth_credentials(location = NULL)
}
\arguments{
\item{location}{file path where token is located}
}
\value{
named listed of override configuration values
}
\description{
Load OAuth Credentials
}
