% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{object_http}
\alias{object_http}
\title{HTTP requests for Object Stores}
\usage{
object_http(
  verb = "GET",
  url,
  datasource_type,
  config = list(),
  headers = list(),
  request_body = "",
  write_disk = NULL
)
}
\arguments{
\item{verb}{A character string containing an HTTP verb, defaulting
to \dQuote{GET}.}

\item{url}{Signed URL to download object at.}

\item{datasource_type}{Type of datasource the object is stored in.}

\item{config}{A list of config values for the REST call.}

\item{headers}{A list of request headers for the REST call.}

\item{request_body}{A character string containing request body data.}

\item{write_disk}{An argument like \code{\link[httr]{write_disk}} to
write the result directly to disk.}
}
\value{
a \code{\link[httr]{response}} object.
}
\description{
HTTP requests for Object Stores
}
\details{
This is mostly for internal use by object store APIs.
}
