% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore_http.R
\name{objectGET}
\alias{objectGET}
\title{GET HTTP requests for Object Stores}
\usage{
objectGET(url, datasource_type, write_disk = NULL)
}
\arguments{
\item{url}{Signed URL to download object at.}

\item{datasource_type}{Type of datasource the object is stored in.}

\item{write_disk}{An argument like \code{\link[httr]{write_disk}} to write the result directly to disk.}
}
\value{
a \code{\link[httr]{response}} object.
}
\description{
GET HTTP requests for Object Stores
}
\details{
This is mostly for internal use by object store APIs.
}
