% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query}
\alias{query}
\title{Query a datasource and returns an arrow Table}
\usage{
query(client, datasource, query, override = new.env())
}
\arguments{
\item{client}{\code{domino_data.data_sources.DataSourceClient}, as returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{query}{The query to run against the provided datasource}

\item{override}{An environment with configuration overrides}
}
\value{
An \link[arrow:Table]{arrow::Table}
}
\description{
Query a datasource and returns an arrow Table
}
