% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore_functions.R
\name{save_object}
\alias{save_object}
\title{Save an object from a datasource to a local file}
\usage{
save_object(
  client,
  datasource,
  object,
  file = basename(object),
  override = new.env()
)
}
\arguments{
\item{client}{\code{domino_data.data_sources.DataSourceClient}, as returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{file}{File path to save object at. Defaults to the object base name.}

\item{override}{An environment with configuration overrides}
}
\value{
Raw vector representation of the object
}
\description{
Save an object from a datasource to a local file
}
