% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore_functions.R
\name{upload_object}
\alias{upload_object}
\title{Upload a file to a datasource}
\usage{
upload_object(client, datasource, object, file, override = new.env())
}
\arguments{
\item{client}{\code{domino_data.data_sources.DataSourceClient}, as returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{file}{File path to upload..}

\item{override}{An environment with configuration overrides}
}
\value{
Raw vector representation of the object
}
\description{
Upload a file to a datasource
}
