% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{datasource_client}
\alias{datasource_client}
\title{Create a client to Domino datasources}
\usage{
datasource_client(api_key = NULL, token_file = NULL)
}
\arguments{
\item{api_key}{string key to override the environment variable}

\item{token_file}{location of file to read token from}
}
\value{
A \code{domino_data.data_sources.DataSourceClient}.
}
\description{
Create a client to Domino datasources
}
