% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.R
\name{get_object}
\alias{get_object}
\title{Retrieve an object from a datasource}
\usage{
get_object(client, datasource, object, as = "raw", override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{as}{Passed through to \code{\link[httr]{content}}}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Raw vector representation of the object
}
\description{
Retrieve an object from a datasource
}
