% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_domino_data_install}
\alias{py_domino_data_install}
\title{Install domino_data Python package}
\usage{
py_domino_data_install()
}
\value{
\code{TRUE} if installation successful, otherwise \code{FALSE}.
}
\description{
Install domino_data Python package
}
\examples{
py_domino_data_install()
}
