% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_domino_data_install}
\alias{py_domino_data_install}
\title{Install domino_data Python package}
\usage{
py_domino_data_install(version)
}
\arguments{
\item{version}{Version of the domino_data package to install.}
}
\value{
\code{TRUE} if installation was successful, \code{FALSE} otherwise.
}
\description{
Install domino_data Python package
}
