% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{py_select_interpreter}
\alias{py_select_interpreter}
\title{Tell {reticulate} to use Python Conda version available on Domino Data Lab.}
\usage{
py_select_interpreter()
}
\value{
\code{TRUE} if a python binary was bound to reticulate, \code{FALSE} otherwise
}
\description{
Provide other options in case the package is used for local development.
}
