% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.R
\name{save_object}
\alias{save_object}
\title{Save an object from a datasource to a local file}
\usage{
save_object(
  client,
  datasource,
  object,
  file = basename(object),
  override = list()
)
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{file}{File path to save object at. Defaults to the object base name.}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Raw vector representation of the object
}
\description{
Save an object from a datasource to a local file
}
