% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstore.R
\name{upload_object}
\alias{upload_object}
\title{Upload a file to a datasource}
\usage{
upload_object(client, datasource, object, file, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{object}{The object to retrieve}

\item{file}{File path to upload..}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Raw vector representation of the object
}
\description{
Upload a file to a datasource
}
